requirements.txt
torch>=2.5.1
torchvision>=0.20.1
numpy>=1.24.4
tqdm>=4.66.1
hydra-core>=1.3.2
iopath>=0.1.10
pillow>=9.4.0
segmentation_models_pytorch==0.4.0
scipy==1.15.1
opencv-python==4.11.0.86